Assignment_2/server/pyproject.toml
[project]
name = "coding-interview-server"
version = "0.1.0"
description = "FastAPI + Socket.IO backend for collaborative code interviews"
authors = [{ name = "AI Dev Tools Zoomcamp" }]
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.23.2",
    "python-socketio[asgi]>=5.11.1",
    "python-engineio>=4.9.1",
    "httpx>=0.26.0",
    "pytest>=8.2.0",
    "pytest-asyncio>=0.21.1",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.uv]
dev-dependencies = [
    "ruff>=0.5.0",
]
